+++ src/structure/collection/linear/list/singly.rs
@@ -1497,7 +1497,7 @@ mod test {
             }
 
             fn size_hint(&self) -> (usize, Option<usize>) {
-                (isize::MAX as usize, Some(isize::MAX as usize))
+                (usize::MAX, Some(usize::MAX))
             }
         }
